♻️ refactor(auth): reorganize login views and fix ApiExtensions
- Move LoginView to Login/CredentialsLoginView for better organization
- Create new modular LoginView as main entry point
- Extract WebLoginView into separate component
- Add KiaTextFieldStyle for consistent text field styling
- Remove unused os.log import from ApiExtensions
- Update AppDelegate with proper view initialization

This refactoring improves code organization by separating login concerns into dedicated components while maintaining the automatic token refresh functionality.
