+++ KiaMaps/App/Login/CredentialsLoginView.swift
@@ -1,62 +1,14 @@
 //
-//  LoginView.swift
+//  CredentialsLoginView.swift
 //  KiaMaps
 //
-//  Created by Lukas Foldyna on 29.05.2024.
-//  Copyright © 2024 Lukas Foldyna. All rights reserved.
+//  Created by Lukáš Foldýna on 1/9/25.
+//  Copyright © 2025 Apple. All rights reserved.
 //
 
 import SwiftUI
-import WebKit
-import os.log
-// Uncomment when package is added:
-// import RecaptchaEnterprise
 
-struct LoginWebView: UIViewRepresentable {
-    let url: URL?
-    let navigationDelegate: WKNavigationDelegate
-
-    func makeUIView(context: Context) -> WKWebView {
-        let configuration = WKWebViewConfiguration()
-        configuration.websiteDataStore = .nonPersistent()
-        configuration.applicationNameForUserAgent = "15E148_CCS_APP_iOS"
-        let webView = WKWebView(frame: .zero, configuration: configuration)
-        webView.isInspectable = true
-        webView.navigationDelegate = navigationDelegate
-        return webView
-    }
-
-    func updateUIView(_ webView: WKWebView, context: Context) {
-        guard let url = url else { return }
-        let request = URLRequest(url: url)
-        webView.load(request)
-    }
-}
-
-class LoginViewWebDelegate: NSObject, WKNavigationDelegate {
-    typealias Callback = (_ code: String, _ state: String, _ loginSuccess: Bool) -> Void
-
-    let api: Api
-    var callback: Callback?
-
-    init(api: Api, callback: Callback?) {
-        self.api = api
-        self.callback = callback
-    }
-
-    func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction) async -> WKNavigationActionPolicy {
-        if let url = navigationAction.request.url, url.path == "/api/v1/user/oauth2/redirect" {
-            guard let extract = try? api.extractAuthorizationCode(from: url) else { return .cancel }
-            callback?(extract.code, extract.state, extract.loginSuccess)
-            return .cancel
-        } else {
-            print(navigationAction.request.url?.path ?? "no path")
-            return .allow
-        }
-    }
-}
-
-struct LoginView: View {
+struct CredentialsLoginView: View {
     @State private var username: String = ""
     @State private var password: String = ""
     @State private var isLoading: Bool = false
@@ -66,10 +18,9 @@ struct LoginView: View {
     @State private var showUsernameError: Bool = false
     @State private var recaptchaToken: String? = nil
     @State private var isVerifyingRecaptcha: Bool = false
-    @State private var showLoginView: Bool = true
 
     @FocusState private var focusedField: Field?
-    
+
     enum Field {
         case username
         case password
@@ -78,16 +29,14 @@ struct LoginView: View {
     let configuration: AppConfiguration.Type
     let api: Api
     let onLoginSuccess: (AuthorizationData) -> Void
-    let delegateObject: LoginViewWebDelegate
 
     init(configuration: AppConfiguration.Type, onLoginSuccess: @escaping (AuthorizationData) -> Void) {
         let api = Api(configuration: configuration.apiConfiguration, rsaService: .init())
         self.configuration = configuration
         self.api = api
         self.onLoginSuccess = onLoginSuccess
-        self.delegateObject = LoginViewWebDelegate(api: api, callback: nil)
     }
-    
+
     var body: some View {
         ScrollView {
             VStack(spacing: KiaDesign.Spacing.large) {
@@ -221,19 +170,6 @@ struct LoginView: View {
                 .padding(.bottom, KiaDesign.Spacing.large)
             }
         }
-        .sheet(isPresented: $showLoginView, content: {
-            NavigationView {
-                Group {
-                    LoginWebView(url: try? api.webLoginUrl(), navigationDelegate: delegateObject)
-                }
-                .ignoresSafeArea()
-                .navigationTitle("Login")
-                .navigationBarTitleDisplayMode(.inline)
-            }
-        })
-        .scrollDismissesKeyboard(.interactively)
-        .background(KiaDesign.Colors.background)
-        .navigationTitle("Login to \(configuration.apiConfiguration.brandName)")
         .onAppear {
             loadSavedCredentials()
             // Auto-focus username field if no saved credentials
@@ -242,22 +178,18 @@ struct LoginView: View {
                     focusedField = .username
                 }
             }
-            self.delegateObject.callback = { code, _, success in
-                guard success else { return }
-                self.finishLogin(code: code)
-            }
         }
     }
 
     private func performLogin() async {
         // Validate email before attempting login
         validateUsername(username)
-        
+
         // Don't proceed if there are validation errors
         guard !showUsernameError else {
             return
         }
-        
+
         // Hide previous errors
         withAnimation(.easeInOut(duration: 0.3)) {
             showError = false
@@ -266,14 +198,14 @@ struct LoginView: View {
             // Set loading state
             isLoading = true
         }
-        
+
         do {
             // Login with API (including reCAPTCHA token)
             let authorizationData = try await api.login(username: username, password: password, recaptchaToken: "")
 
             // Store credentials and authorization data
             storeCredentials()
-            
+
             // Call success callback with authorization data
             await MainActor.run {
                 isLoading = false
@@ -298,70 +230,35 @@ struct LoginView: View {
         }
     }
 
-    private func finishLogin(code: String) {
-        showLoginView = false
-        errorMessage = ""
-        isLoading = true
-        Task {
-            do {
-                // Login to get tokens
-                let authorizationData = try await api.login(authorizationCode: code)
 
-                // Store credentials and authorization data
-                storeCredentials()
-
-                // Call success callback with authorization data
-                await MainActor.run {
-                    isLoading = false
-                    onLoginSuccess(authorizationData)
-                }
-            } catch let apiError as ApiError {
-                await MainActor.run {
-                    isLoading = false
-                    errorMessage = apiError.localizedDescription
-                    withAnimation(.easeInOut(duration: 0.3)) {
-                        showError = true
-                    }
-                }
-            } catch {
-                await MainActor.run {
-                    isLoading = false
-                    errorMessage = "An unexpected error occurred. Please try again."
-                    withAnimation(.easeInOut(duration: 0.3)) {
-                        showError = true
-                    }
-                }
-            }
-        }
-    }
 
     private func storeCredentials() {
         // Store credentials securely in keychain
         let credentials = LoginCredentials(username: username, password: password)
         LoginCredentialManager.store(credentials: credentials)
     }
-    
+
     private func loadSavedCredentials() {
         if let savedCredentials = LoginCredentialManager.retrieveCredentials() {
             username = savedCredentials.username
             password = savedCredentials.password
         }
     }
-    
+
     private func validateUsername(_ email: String) {
         // Clear previous errors first
         withAnimation(.easeInOut(duration: 0.2)) {
             showUsernameError = false
             usernameError = ""
         }
-        
+
         // Don't validate empty field (let required field validation handle it)
         guard !email.isEmpty else { return }
-        
+
         // Email validation regex
         let emailRegex = #"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$"#
         let emailPredicate = NSPredicate(format: "SELF MATCHES %@", emailRegex)
-        
+
         if !emailPredicate.evaluate(with: email) {
             withAnimation(.easeInOut(duration: 0.2)) {
                 usernameError = "Please enter a valid email address"
@@ -371,34 +268,10 @@ struct LoginView: View {
     }
 }
 
-// MARK: - Custom Text Field Style
-
-struct KiaTextFieldStyle: TextFieldStyle {
-    let hasError: Bool
-    
-    init(hasError: Bool = false) {
-        self.hasError = hasError
-    }
-    
-    func _body(configuration: TextField<_Label>) -> some View {
-        configuration
-            .padding(KiaDesign.Spacing.medium)
-            .background(
-                RoundedRectangle(cornerRadius: 12)
-                    .fill(KiaDesign.Colors.cardBackground)
-                    .stroke(
-                        hasError ? KiaDesign.Colors.error : KiaDesign.Colors.textTertiary.opacity(0.2), 
-                        lineWidth: hasError ? 2 : 1
-                    )
-            )
-            .font(KiaDesign.Typography.body)
-    }
-}
-
 // MARK: - Preview
 
 #Preview {
-    LoginView(configuration: AppConfiguration.self) { authData in
+    CredentialsLoginView(configuration: AppConfiguration.self) { authData in
         logInfo("Login successful for user", category: .auth)
     }
 }
